# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Configurations API",
        "description": "API for managing cars, configurations, options and prices",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/options": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Get all options",
                "operationId": "7bdf34595bba50876aa6f482b829e265",
                "responses": {
                    "200": {
                        "description": "List of all options",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Options"
                ],
                "summary": "Create a new option",
                "operationId": "e7ff098bfb547f5b34d9eab1c96bc21f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Climate Control"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Option created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/options/{id}": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Get a single option",
                "operationId": "07832a1d1cbae465b8ad194d537bb2ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the option to get",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Option details",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Option not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Options"
                ],
                "summary": "Update an option",
                "operationId": "b6b6224491b9ec8bd05ac56f076cb040",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the option to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Climate Control"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Option updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Option not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Options"
                ],
                "summary": "Delete an option",
                "operationId": "4f0a5219420f26b1fe8fc3948cd92be6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the option to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Option deleted successfully"
                    },
                    "404": {
                        "description": "Option not found"
                    }
                }
            }
        },
...
    },
...
}